[PE-7043][PE-7096] Some refactoring for buy/sell keyboard avoidance (#13182)
### Description
It's not a full fix, but it makes this slightly better. I can follow up
with a more ambitious plan to track the currently focused input and make
sure it stays in view when the keyboard is shown.

* Remove usage of `FixedFooter` in favor of a less-nested element that
stays hidden behind keyboard on iOS (but still pinned to bottom) and is
correctly positioned above keyboard on Android (there was a gap below
the element when rendered there)
* Uses our `KeyboardAvodingView` instead of manually adding padding
based on keyboard visibility. I updated that view to support the
'padding' mode where it adjusts bottom padding instead of translating
the content. The reason I want to use `KeyboardAvoidingView` is that it
sets the new padding before showing/hiding instead of us doing a render
afterwards that feels janky
* Note: For Android, since KAV doesn't do anything (because we don't
watch the right events there), I chose to add a fixed padding amount to
the bottom of the scrollview content. This will ensure that you can
scroll the view far enough to see everything if the footer makes its way
above the content due to the keyboard showing.
* Fixed a bug where switching tabs while focused would not blur the
input, causing keyboard to stay visible. This due to our usage of
`keyboardShouldPersistTaps='handled'` (good idea for most of this view,
but the segment controls do handle the press, so keyboard will stay
unless we manually dismiss it)
* Tiny bonus fix for PE-7096 to remove an extra divider in the coins
section

### How Has This Been Tested?
local prod build on physical iOS and android devices. Did a lot of
focusing and blurring in the various tabs of the buy/sell flow.