[PE-7026] Re-enable dismissing of toasts on Android (#13188)
### Description
This change: https://github.com/AudiusProject/apps/pull/10147
combined with this change:
https://github.com/AudiusProject/audius-client/pull/3838
causes toasts to re-trigger when you load the BuySellModalScreen since
we don't actually dismiss them on Android.

I was unable to reproduce the original issue that this was intended to
fix (flag and hide a comment and toasts break on Android), so I'm
proposing we remove this conditional.

If it feels too dangerous, I can try to rig up an alternate solution
where we set some kind of `dismissed` flag on toasts and then don't show
them on mount of a new Toast component.

### How Has This Been Tested?
Local prod build on Pixel 5 and Pixel 7, triggered various toasts and
checked that they don't show when opening the `BuySellModalScreen`.