[PE-7110] Fix PlainButton link wrapping (#13154)
### Description
See original description in #13153 

* Fixes PlainButton to be asChild-aware and skip the wrapper span
* Converts color styling to not use a CSS variable. The style reset for
links ends up jumping in front of it in priority. `Button` doesn't use a
variable (probably for the same reason?). So I updated it to just apply
`color` and `fill` directly.

### How Has This Been Tested?
Viewed all instances of `asChild` usage in `PlainButton` on local app
against staging and verified they rendered correctly.