[PE-7043] Add KeyboardAwareScrollView to buy/sell flow (#13195)
### Description
This leverages
[react-native-keyboard-controller](https://kirillzyusko.github.io/react-native-keyboard-controller/)
to control scroll position on the ScrollView in the buy/sell flow. I've
tested it out on physical iOS and Android devices and it seems pretty
dang smooth. Generally could be a good replacement for all of our
'keyboard avoiding' type components. But we can do that one-by-one to
reduce risk.

### How Has This Been Tested?
Staging apps on iOS and Android hardware devices. Went through the
buy/sell screens, including convert, and made sure the focused input was
moved above the keyboard when it is shown.