Fix android notification crashes from bad env vars (#13215)
### Description
Root cause are the changes to notifs from
https://github.com/AudiusProject/apps/pull/12897
Not sure if this crashes iOS as well, but tip received notifs cause a
force close on my device.
`services/env` appears to map to the web folder. I searched the mobile
codebase and noticed we usually import this as `app/services/env`

### How Has This Been Tested?
Pixel 7 Pro prod build. Opened my notifs and no more crashes!