Fix double reaction bug (#13329)
Indexing was getting stalled from the previous fix because the primary
key would already exist.

This change reverts #13282 and fixes it a different way by moving the
check for existing deleted comments to the validation logic, and adding
the code to handle that case in the reaction creation logic.

Also adds commands for `audius-cmd` and tested locally.

**NOTE**: Re-liking a comment will _not_ trigger a new notification.
Made a call that we probably don't want to spam notifs every time it
toggles.