[PE-7263] Fix external wallets failing to disconnect (#13342)
### Description
Co-author: @rickyrombo branch `mjp-ass-wallet`

Bug was that there was a duplicate row for the same wallet address,
associated with a different user, but with `is_delete = True` which
caused `existing_records` to pick up that row instead.

Added logic to include all records for the given wallet address, then
cycle through them and make sure we delete the right row with the
correct user ID.

There's a bigger convo around whether we should be deleting rows
outright, or marking them with `is_delete = True` like we do for other
operations.

### How Has This Been Tested?

Added a row with `is_delete = True` to the existing wallet removal test.
Also added a new test case to confirm that associating a wallet that's
already associated with a different user results in the old row getting
deleted and the new association getting created. Also added a row with
`is_delete = True` for this case.