Add DOM test output on web-test failures (#13383)
### Context
CoinDetailPage test (that I just added) is flaking sometimes looking for
the creator's avatar image. Haven't determined why but its gotta be
something with our image loading. Not clear to me if its an actual bug
or if its a test quirk.
This test was flaking super intermittently and it was hard to repro
locally, so I had the idea to get some DOM output on failure.

### Description

- Wraps vitest's `it()` fn to automatically output to test-output on
failure (this folder is gitignored)
- Comment out the flaky piece of the CoinDetailPageTest for now. 
- Originally I was planning to leave this in to get some DOM data, but
while testing this change locally I managed to get a failure repro and
it spit out DOM for me so I have a reference now.

### How Has This Been Tested?

npx vitest