Timeout image loading in case of hanging connections (#13409)
### Description

Some nodes might be down/unreachable in a way that causes connections to
hang open indefinitely. This creates a default 5 second timeout for
loading images before timing out to allow the fetcher to move on to the
next mirror.

### How Has This Been Tested?

Ran local web stack against a prod track. The track's artwork was stored
on a node that creates hanging connections. The timeout triggered after
the 5 second default and the artwork loaded quickly once the fetcher
was able to try the next working mirror.