Use new validators endpoint in healthz to fetch nodes (#13462)
### Description

Healthz uses an api endpoint that returns an outdated list of registered
nodes. This switches it to use the newly created endpoint, keeping info
up-to-date.

This is helpful because healthz consolidates a bunch of useful
information about the state of storage on various machines. This
information may be ported to the go-openaudio console at some point, but
for now this is the easiest and most accessible place for it.

This also removes additional unused UI components such as the `dev` tab.

### How Has This Been Tested?

Ran locally and confirmed it loads data from the new api endpoint.